+++ src/main/java/com/doctor/vampiricageing/config/WerewolvesAgeingConfig.java
@@ -21,7 +21,7 @@ public class WerewolvesAgeingConfig {
     public static final ForgeConfigSpec.DoubleValue biteSaturation;
     public static final ForgeConfigSpec.ConfigValue<List<? extends Integer>> devouredForNextAge;
     public static final ForgeConfigSpec.ConfigValue<List<? extends Integer>> nourishmentMultipliers;
-    public static final ForgeConfigSpec.ConfigValue<List<? extends Float>> biteDamageMultiplier;
+    public static final ForgeConfigSpec.ConfigValue<List<? extends Double>> biteDamageMultiplier;
     public static final ForgeConfigSpec.ConfigValue<List<? extends Float>> silverOilDamageMultiplier;
     public static final ForgeConfigSpec.ConfigValue<List<? extends Float>> formTimeMultiplier;
     public static final ForgeConfigSpec.ConfigValue<List<? extends Float>> healonBiteAmount;
@@ -41,7 +41,7 @@ public class WerewolvesAgeingConfig {
         greaterDevourWorth = COMMON_BUILDER.comment("How much a greater devour is worth. These are things like advanced faction mobs").defineInRange("greaterDevourWorth", 4, 0, 99);
         exquisiteDevourWorth = COMMON_BUILDER.comment("How much a exquisite devour is worth. These are things like vampire barons").defineInRange("exquisiteDevourWorth", 8, 0, 99);
         maxHealthIncrease = COMMON_BUILDER.comment("Max Health Increase for each rank. This is addition, not multiplier based").defineList("maxHealthIncrease", Arrays.asList(0D, 2D, 2D, 4D, 4D, 6D), t -> true);
-        biteDamageMultiplier = COMMON_BUILDER.comment("How much each rank multiplies bite damage. Bite damage multiplier is strange and values of 0 should be treated as 1, values like 0.2 should be treated as a 1.2 multiplier").defineList("biteDamageMultiplier", Arrays.asList(0f, 0f, 0f, 0.125f, 0.25f, 0.5f), it -> true);
+        biteDamageMultiplier = COMMON_BUILDER.comment("How much each rank multiplies bite damage. Bite damage multiplier is strange and values of 0 should be treated as 1, values like 0.2 should be treated as a 1.2 multiplier").defineList("biteDamageMultiplier", Arrays.asList(0d, 0d, 0d, 0.125d, 0.25d, 0.5d), it -> true);
         silverOilDamageMultiplier = COMMON_BUILDER.comment("How much each rank multiplies damages from Silver Oil. Values are set to 1 by default as to disable it by default.").defineList("silverOilMultiplier", Arrays.asList(1f, 1f, 1f, 1f, 1f, 1f), it -> true);
         healonBiteAmount = COMMON_BUILDER.comment("How much the player is healed for biting an entity based on rank").defineList("healonBiteAmount", Arrays.asList(0f, 0f, 0f, 1f, 2f, 2f), it -> true);
         formTimeMultiplier = COMMON_BUILDER.comment("How much the duration of time a player can stay in werewolf form is multiplied by").defineList("formTimeMultiplier", Arrays.asList(1f, 1.5f, 2f, 3f, 4f, 5f), it -> true);
